Rewrite functionality into a plugin
Having the functionality as a storage was problematic for multiple
reasons:

* UploadedFile#url always returned the Imgix URL, it wasn't possible
  to return the S3 URL unless you drop two levels down to the S3 storage
  level.

* The storage delegator was very hacky and brittle. It's just a wrong kind
  of implementation.

Instead, we rewrite the functionality into a plugin. Instead of
overriding UploadedFile#url (which would make it intrusive), the plugin
adds a new UploadedFile#imgix_url method which returns the Imgix URL.
This is consistent with how other plugins in Shrine such as
download_endpoint and derivation_endpoint work.

We also make purging on delete opt-in, because Imgix says it's only
necessary if we're replacing files or have a need to expire cache, so
there is no reason to make an additional HTTP request by default.

We also make including storage prefix opt-out instead of opt-in, because
I expect it will be more common to configure the Imgix source without a
path prefix.